+++ _src/interfaces/serviceRequestQueueEntry.ts
@@ -51,7 +51,12 @@ export interface ServiceRequestQueueEntry {
 	interval?: number;
 
 	/**
-	 * date after which no further request shall be executed
+	 * alternative for interval - it is possible to configure a recurring execution by a cron-like syntax
+	 */
+	recurring?: CronConfig;
+
+	/**
+	 * date after which no further request shall be executed, in case interval or recurring is set
 	 */
 	endOfInterval?: Date;
 
@@ -59,6 +64,30 @@ export interface ServiceRequestQueueEntry {
 	//  * space for individual data e.g. to use it in the isResponsible Function or whatever
 	//  */
 	// params?: TParams;
+}
+
+/**
+ * each property can be a number, comma-separated numbers or a '*'
+ * e.g.S
+ * 	month: '1,4': each January an April
+ * 	dayOfWeek: '1': each Monday
+ * 	year: '*': each year
+ * 	hour: '10': always at hour 10
+ * 	minute: '5': always at minute 5
+ * 	second: '0': always on second 0
+ *
+ * 	all this together means: each year on each Monday in January and April at 10:05:00
+ *
+ * 	skipping any property is the same as setting it to '*'
+ */
+export interface CronConfig {
+	second?: string;
+	minute?: string;
+	hour?: string;
+	dayOfMonth?: string;
+	month?: string;
+	dayOfWeek?: string;
+	year?: string;
 }
 
 export interface ServiceRequestQueueEntryQueryConditions {